requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Cogs==0.4.4
Deprecated==1.2.12
discord-py-slash-command==1.2.2
discord.py==1.7.2
discord_components==1.0.0
idna==2.10
lxml==4.6.3
multidict==5.1.0
pycodestyle==2.7.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyxDamerauLevenshtein==1.7.0
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wrapt==1.12.1
yarl==1.6.3
